configuration spring security

+++ src/main/java/co/edu/unbosque/ciclo3back/dao/UsuariosDAO.java
@@ -1,8 +1,13 @@
 package co.edu.unbosque.ciclo3back.dao;
 
+import javax.persistence.EntityManager;
+
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import co.edu.unbosque.ciclo3back.model.Usuarios;
 
-public interface UsuariosDAO extends JpaRepository<Usuarios, Integer> {
+public interface UsuariosDAO extends JpaRepository<Usuarios, Long> {
 
+	
+	
 }
